requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiostream==0.5.2
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
astor==0.8.1
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==6.1.0
blinker==1.4
bokeh==2.2.2
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.2
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.18
cliapp==1.20180812.1
click==8.0.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmdtest==0.32+git
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.5.14
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.7.0
distro-info==1.1+ubuntu0.1
docopt==0.6.2
duckdb==0.8.1
duplicity==0.8.21
elevenlabs==0.2.26
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.104.1
fasteners==0.14.1
fastjsonschema==2.18.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.4.0
fs==2.4.12
fsspec==2023.10.0
ftfy==6.1.3
future==0.18.2
gast==0.5.2
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
gyp==0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.3
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keyring==23.5.0
kiwisolver==1.3.2
kubernetes==28.1.0
langchain==0.0.336
langsmith==0.0.66
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
llama-index==0.9.4
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.2.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.16.3
openai==1.3.5
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandasai==1.4.6
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipdeptree==2.13.1
pipreqs==0.4.13
platformdirs==3.11.0
ply==3.11
portalocker==2.8.2
posthog==3.0.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.0
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.11
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypdf==3.17.1
PyPika==0.48.9
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-json-logger==2.0.7
pythran==0.10.0
pyttsx3==2.90
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.1
qdrant-client==1.6.9
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.4
rsa==4.9
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.3
SecretStorage==3.3.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
stack-data==0.6.3
starlette==0.27.0
streamlit==1.29.0
streamlit-bokeh-events==0.1.2
streamlit-chat==0.1.1
sympy==1.9
systemd-python==234
tenacity==8.2.3
terminado==0.17.1
tifffile==2023.9.26
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.35.2
triton==2.1.0
ttystatus==0.38
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
urllib3==1.26.18
usb-creator==0.3.7
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
widgetsnbextension==4.0.9
wrapt==1.16.0
xdg==5
xkit==0.0.0
yarg==0.1.9
yarl==1.9.2
zipp==1.0.0
FlagEmbedding